+++ test/unit/fixtures/bin-config.js
@@ -1,0 +1,3 @@
+export default {
+  breaks: true
+};
